Moving Exec(...) test to 1.9 build as NamedValueChecker is only supported since 1.9
