+++ expectations_before_go18.go
@@ -3,6 +3,7 @@
 package sqlmock
 
 import (
+	"database/sql/driver"
 	"fmt"
 	"reflect"
 )
@@ -37,6 +38,10 @@ func (e *queryBasedExpectation) argsMatches(args []namedValue) error {
 		darg, err := e.converter.ConvertValue(dval)
 		if err != nil {
 			return fmt.Errorf("could not convert %d argument %T - %+v to driver value: %s", k, e.args[k], e.args[k], err)
+		}
+
+		if !driver.IsValue(darg) {
+			return fmt.Errorf("argument %d: non-subset type %T returned from Value", k, darg)
 		}
 
 		if !reflect.DeepEqual(darg, v.Value) {
